move function into pillowtop.es_utils
