+++ corehq/ex-submodules/pillowtop/es_utils.py
@@ -1,3 +1,4 @@
+from dimagi.ext import jsonobject
 from collections import namedtuple
 from copy import copy
 from datetime import datetime
@@ -127,3 +128,20 @@ def doc_exists(pillow, doc_id_or_dict):
 def get_all_elasticsearch_pillow_classes():
     from pillowtop.listener import AliasedElasticPillow
     return filter(lambda x: issubclass(x, AliasedElasticPillow), get_all_pillow_classes())
+
+
+class ElasticsearchIndexInfo(jsonobject.JsonObject):
+    index = jsonobject.StringProperty(required=True)
+    alias = jsonobject.StringProperty()
+
+
+def get_all_expected_es_indices():
+    """
+    Get all expected elasticsearch indices according to the currently running code
+    """
+    seen_indices = set()
+    pillows = get_all_elasticsearch_pillow_classes()
+    for pillow in pillows:
+        assert pillow.es_index not in seen_indices
+        yield ElasticsearchIndexInfo(index=pillow.es_index, alias=pillow.es_alias)
+        seen_indices.add(pillow.es_index)
